Add quotes to descriptions (website_issue.yml)
I am aware of another syntax error (very vague) but this seems like a
useful thing to do first. It might solve the problem but I am not sure
as I do not know YAML (though syntax highlighting seems to not show any
errors) and the error is too vague to be sure.